requirements.txt
# requirements.txt
fastapi==0.111.0
pydantic==2.7.1
uvicorn==0.29.0
ollama==0.1.7
PyMuPDF==1.23.21
python-docx==1.1.0
python-multipart==0.0.9
